requirements.txt
asgiref==3.7.2
Django==4.2.7
django-configurations==2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pip==23.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.2
whitenoise==6.6.0
#uwsgi
